+++ src/test/java/com/epam/taskgym/services/TrainerServiceTest.java
@@ -1,4 +1,4 @@
-package com.epam.taskgym;
+package com.epam.taskgym.services;
 
 import com.epam.taskgym.models.TrainerDTO;
 import com.epam.taskgym.entity.Trainer;
