+++ src/test/java/com/epam/taskgym/services/UserServiceTest.java
@@ -1,4 +1,4 @@
-package com.epam.taskgym;
+package com.epam.taskgym.services;
 
 import com.epam.taskgym.entity.User;
 import com.epam.taskgym.exception.MissingAttributes;
